Development of a carrousel to display images on portfolio
Currently I am developing a carrousel using a json feed, to make it
reusable for different projects. The basic structure of of the js file
is an input which is the path to the location of the specific json feed.
The json feed has the data, where images are located, but also the text
commenting the images.